Change everything to Int, make it so that the values aren't more than an octave apart, and not 6 or 11 half steps apart.

+++ Music Generator/Models/MusicBox.swift
@@ -45,7 +45,7 @@ class MusicBox {
         }
         
         let event = MIKMIDINoteEvent(timeStamp: location,
-                                     note: value,
+                                     note: UInt8(value),
                                      velocity: 65,
                                      duration: Float32(duration.value),
                                      channel: 0)
